Toned down the color printing, and all the printing in general. Still prints where necessary with a little organization.
Gave me the better idea that win-backbone shouldn't force the module to declare it's type during emit calls. Instead, win-backbone should create an emit wrapper for each module that the module uses for emit.

In the beginning, modules call winBackbone.getEmitter(self); This will return the object to call for passing functions. That way, it can wrap module properties specifically in the future, and pull any changes without issues.